bsides-2019/consulta_multas/requirements.txt
django==2.0.2
mysqlclient
djangorestframework
djangorestframework-jwt
